+++ cfgov/v1/models/__init__.py
@@ -1,6 +1,7 @@
 from django.conf import settings
 from .base import *
 from .events import *
+from .atoms import *
 from .molecules import *
 from .organisms import *
 from .snippets import *